#1071 Re-structure component provider implementation
