+++ hartshorn-config/src/main/java/org/dockbox/hartshorn/config/ConfigurationURIContextList.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019-2023 the original author or authors.
+ * Copyright 2019-2024 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,12 +19,13 @@ package org.dockbox.hartshorn.config;
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 
+import org.dockbox.hartshorn.context.ContextIdentity;
 import org.dockbox.hartshorn.context.ContextKey;
 import org.dockbox.hartshorn.context.DefaultProvisionContext;
 
 public class ConfigurationURIContextList extends DefaultProvisionContext {
 
-    public static final ContextKey<ConfigurationURIContextList> CONTEXT_KEY = ContextKey.builder(ConfigurationURIContextList.class)
+    public static final ContextIdentity<ConfigurationURIContextList> CONTEXT_KEY = ContextKey.builder(ConfigurationURIContextList.class)
             .fallback(ConfigurationURIContextList::new)
             .build();
