+++ hartshorn-core/src/main/java/org/dockbox/hartshorn/component/ComponentProviderPostProcessor.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2019-2024 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.dockbox.hartshorn.component;
+
+import org.dockbox.hartshorn.inject.ComponentRequestContext;
+import org.dockbox.hartshorn.inject.ObjectContainer;
+import org.dockbox.hartshorn.util.ApplicationException;
+
+public interface ComponentProviderPostProcessor {
+
+    <T> T processInstance(
+            ComponentKey<T> componentKey,
+            ObjectContainer<T> objectContainer,
+            T instance,
+            ComponentRequestContext requestContext
+    ) throws ApplicationException;
+}
